Update android_emulator.md
Basically just an edit pass, for things that bothered me. I tried to tone down some of the more marketing-oriented phrasing, to match the tone guidelines. I'm concerned that there's too much "how to" material with all the UI commands listed, but I didn't make a pass to remove it. I think we're just supposed to note the capabilities, and not describe how to set them.